Few changes

+++ src/tme/models/chunk.rs
@@ -2,12 +2,12 @@ use super::data_source::DataSource;
 use serde::Deserialize;
 use serde::Serialize;
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, PartialEq)]
 #[serde(rename_all = "lowercase")]
 pub struct Chunk {
     pub data:   DataSource,
-    pub height: i32,
-    pub width:  i32,
-    pub x:      i32,
-    pub y:      i32,
+    pub height: i64,
+    pub width:  i64,
+    pub x:      i64,
+    pub y:      i64,
 }
