+++ src/tme/models/frame.rs
@@ -4,6 +4,7 @@ use serde::Serialize;
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "lowercase")]
 pub struct Frame {
-    pub duration: i32,
-    pub tiled_id: i32,
+    pub duration: i64,
+    #[serde(rename = "tiledid")]
+    pub tiled_id: i64,
 }
