+++ src/tme/models/grid.rs
@@ -6,7 +6,7 @@ use super::orientation::Orientation;
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "lowercase")]
 pub struct Grid {
-    pub height:      i32,
+    pub height:      i64,
     pub orientation: Orientation,
-    pub width:       i32,
+    pub width:       i64,
 }
