+++ src/tme/models/polygon_object.rs
@@ -4,11 +4,11 @@ use serde::Serialize;
 use super::point::Point;
 use super::property::Property;
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, PartialEq)]
 #[serde(rename_all = "lowercase")]
 pub struct PolygonObject {
     pub height:     f64,
-    pub id:         i32,
+    pub id:         i64,
     pub name:       String,
     pub properties: Option<Vec<Property>>,
     pub polygon:    Vec<Point>,
