+++ src/tme/models/rectangle_object.rs
@@ -3,11 +3,11 @@ use serde::Serialize;
 
 use super::property::Property;
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, PartialEq)]
 #[serde(rename_all = "lowercase")]
 pub struct RectangleObject {
     pub height:     f64,
-    pub id:         i32,
+    pub id:         i64,
     pub name:       String,
     pub properties: Option<Vec<Property>>,
     pub rotation:   f64,
