+++ src/tme/models/terrain.rs
@@ -8,5 +8,5 @@ use super::property::Property;
 pub struct Terrain {
     pub name:       String,
     pub properties: Option<Vec<Property>>,
-    pub tile:       i32,
+    pub tile:       i64,
 }
