+++ src/tme/models/text.rs
@@ -3,24 +3,29 @@ use serde::Serialize;
 
 use crate::tme::color::Color;
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, PartialEq)]
 #[serde(rename_all = "lowercase")]
 pub struct Text {
     pub bold:        bool,
     pub color:       Color,
+    #[serde(rename = "fontfamily")]
     pub font_family: String,
+    #[serde(rename = "halign")]
     pub h_align:     HorizontalAlign,
     pub italic:      bool,
     pub kerning:     bool,
-    pub pixel_size:  i32,
+    #[serde(rename = "pixelsize")]
+    pub pixel_size:  i64,
+    #[serde(rename = "strikeout")]
     pub strike_out:  bool,
     pub text:        String,
     pub underline:   bool,
+    #[serde(rename = "valign")]
     pub v_align:     VerticalAlign,
     pub wrap:        bool,
 }
 
-#[derive(Debug, Serialize, Deserialize, Copy, Clone)]
+#[derive(Debug, Serialize, Deserialize, Copy, Clone, PartialEq, Eq)]
 #[serde(rename_all = "lowercase")]
 pub enum HorizontalAlign {
     Center,
@@ -29,7 +34,7 @@ pub enum HorizontalAlign {
     Left,
 }
 
-#[derive(Debug, Serialize, Deserialize, Copy, Clone)]
+#[derive(Debug, Serialize, Deserialize, Copy, Clone, PartialEq, Eq)]
 #[serde(rename_all = "lowercase")]
 pub enum VerticalAlign {
     Center,
