+++ src/tme/models/text_object.rs
@@ -4,11 +4,11 @@ use serde::Serialize;
 use super::property::Property;
 use super::text::Text;
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, PartialEq)]
 #[serde(rename_all = "lowercase")]
 pub struct TextObject {
     pub height:     f64,
-    pub id:         i32,
+    pub id:         i64,
     pub name:       String,
     pub properties: Option<Vec<Property>>,
     pub rotation:   f64,
